Update ids to bigint in case we uh run over 2 million speedfests or something. Also remove old .sql files from solution

+++ MWSFDataAccess/Models/IGenericModel.cs
@@ -8,7 +8,7 @@ namespace MWSFDataAccess.Models
 {
     public interface IDatabaseModel
     {
-        int Id { get; set; }
+        ulong Id { get; set; }
         public DateTime? Created { get; set; }
         public DateTime? Modified { get; set; }
     }
